+++ poo/src/main/java/com/ecomerce/poo/model/ProdutoModel.java
@@ -16,13 +16,13 @@ public class ProdutoModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "CDPRODUTO")
-    private int cdProduto;
+    private Long cdProduto;
     @Column(name = "NMPRODUTO")
     private String nmProduto;
     @Column(name = "DSPRODUTO")
     private String dsProduto;
     @Column(name = "VLPRODUTO")
-    private double vlProduto;
+    private Double vlProduto;
     @Column(name = "FLATIVO")
     private String flAtivo;
 }
